concord-server: only admin users should be able to update template aliases

+++ server/impl/src/main/java/com/walmartlabs/concord/server/project/TemplateAliasResourceImpl.java
@@ -5,6 +5,8 @@ import com.walmartlabs.concord.server.api.project.TemplateAliasEntry;
 import com.walmartlabs.concord.server.api.project.TemplateAliasResource;
 import com.walmartlabs.concord.server.api.project.TemplateAliasResponse;
 import com.walmartlabs.concord.server.api.security.Permissions;
+import com.walmartlabs.concord.server.security.UserPrincipal;
+import org.apache.shiro.authz.UnauthorizedException;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.jooq.Configuration;
 import org.sonatype.siesta.Resource;
@@ -27,10 +29,13 @@ public class TemplateAliasResourceImpl extends AbstractDao implements TemplateAl
     @Override
     @RequiresPermissions(Permissions.TEMPLATE_MANAGE)
     public TemplateAliasResponse createOrUpdate(TemplateAliasEntry request) {
+        assertAdmin();
+
         tx(tx -> {
             aliasDao.delete(request.getAlias());
             aliasDao.insert(request.getAlias(), request.getUrl());
         });
+
         return new TemplateAliasResponse();
     }
 
@@ -43,7 +48,15 @@ public class TemplateAliasResourceImpl extends AbstractDao implements TemplateAl
     @Override
     @RequiresPermissions(Permissions.TEMPLATE_MANAGE)
     public TemplateAliasResponse delete(String alias) {
+        assertAdmin();
         aliasDao.delete(alias);
         return new TemplateAliasResponse();
+    }
+
+    private static void assertAdmin() {
+        UserPrincipal p = UserPrincipal.getCurrent();
+        if (!p.isAdmin()) {
+            throw new UnauthorizedException("Not authorized");
+        }
     }
 }
